% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\alias{app_ui}
\alias{app_server}
\alias{app_config}
\alias{app_resources}
\alias{run_app}
\title{Shiny Application}
\usage{
app_ui(request)

app_server(input, output, session)

app_config(
  value,
  config = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  use_parent = TRUE
)

app_resources()

run_app(...)
}
\arguments{
\item{request}{Internal parameter for \code{{shiny}}.}

\item{input, output, session}{Internal parameters for {shiny}.}

\item{value}{Value to retrieve from the config file.}

\item{config}{R_CONFIG_ACTIVE value.}

\item{use_parent}{Logical, scan the parent directory for config file.}

\item{...}{A series of options to be used inside the app.}
}
\description{
The shiny application is developed using the golem framework.
}
\section{Functions}{
\itemize{
\item \code{app_ui}: Application User Interface

\item \code{app_server}: Application Server Logic

\item \code{app_config}: Application Configuration

\item \code{app_resources}: Application Resources

\item \code{run_app}: Run the Application
}}

