% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-tjutils.R
\docType{import}
\name{dev-cicd}
\alias{dev-cicd}
\alias{dev_cicd_test_coverage}
\alias{reexports}
\alias{dev_cicd_test_linting}
\alias{dev_cicd_test_spelling}
\alias{dev_cicd_test_package}
\alias{dev_release_package}
\title{Package CI/CD Functions}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tjutils}{\code{\link[tjutils:dev_cicd]{dev_cicd_test_coverage}}, \code{\link[tjutils:dev_cicd]{dev_cicd_test_linting}}, \code{\link[tjutils:dev_cicd]{dev_cicd_test_package}}, \code{\link[tjutils:dev_cicd]{dev_cicd_test_spelling}}, \code{\link[tjutils:dev_release]{dev_release_package}}}
}}

