% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.something-package.R
\docType{package}
\name{watch.something-package}
\alias{watch.something}
\alias{watch.something-package}
\title{watch.something: Watch Something}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An app to decide what ot watch
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tjpalanca/watch.something}
  \item Report bugs at \url{https://github.com/tjpalanca/watch.something/issues}
}

}
\author{
\strong{Maintainer}: Troy James Palanca \email{mail@tjpalanca.com} (\href{https://orcid.org/0000-0002-0610-9213}{ORCID})

}
\keyword{internal}
